Expose request implicits to Java Api
